% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage_helper_functions.R
\name{apply_output_cutoffs}
\alias{apply_output_cutoffs}
\title{Apply Output Cutoffs}
\usage{
apply_output_cutoffs(linkage_db, algorithm_id, output_df)
}
\arguments{
\item{linkage_db}{A database connection to the linkage metadata.}

\item{algorithm_id}{An iteration number.}

\item{output_df}{The data frame being used during report generation}
}
\description{
The apply_output_cutoffs() function will take in a linkage database connection containing
all the metadata, along with an algorithm ID and the dataframe being used for report generation.
A dataframe is returned which contains the dataset fields being kept (with cutoffs)
}
\examples{
sqlite_file <- file.choose() # Select the '.sqlite' linkage metadata file
linkage_db <- dbConnect(SQLite(), sqlite_file)
algorithm_id <- 1
apply_output_cutoffs(linkage_db, algorithm_id, output_df)
}
