% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage_helper_functions.R
\name{collapse_with_and}
\alias{collapse_with_and}
\title{Collapse With And}
\usage{
collapse_with_and(input)
}
\arguments{
\item{input}{A vector of values to collapse into a singular string}
}
\description{
The collapse_with_and() function will paste a vector of values together, separating all
values with comma except for the last value which will be separated with an "And"
}
\examples{
collapsed_ground_truth <- collapse_with_and(c('PHIN', 'GUID'))

}
