% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage_helper_functions.R
\name{create_extra_parameters_list}
\alias{create_extra_parameters_list}
\title{Create Extra Parameters List}
\usage{
create_extra_parameters_list(
  linkage_output_folder = NULL,
  output_linkage_iterations = FALSE,
  output_unlinked_iteration_pairs = FALSE,
  linkage_report_type = NULL,
  calculate_performance_measures = FALSE,
  data_linker = NULL,
  generate_algorithm_summary = FALSE,
  generate_threshold_plots = FALSE
)
}
\arguments{
\item{linkage_output_folder}{A directory for which linkage output files will be written to.}

\item{output_linkage_iterations}{A TRUE or FALSE value for whether you'd like each iteration to write the linked pairs (UNEDITED) to the output directory.}

\item{output_unlinked_iteration_pairs}{A TRUE or FALSE value for whether you'd like each iteration to write the unlinked pairs (UNEDITED) to the output directory}

\item{linkage_report_type}{1 = No Report, 2 = Intermediate Report, 3 = Final Report.}

\item{calculate_performance_measures}{A TRUE or FALSE value for whether you'd like to calculate and export performance measures from the algorithms being run.}

\item{data_linker}{A single string input for whom performed the data linkage (used for generating a Linkage Quality Report).}

\item{generate_algorithm_summary}{A TRUE or FALSE value for whether you'd like to export a CSV summary of the algorithm that was run.}

\item{generate_threshold_plots}{A TRUE or FALSE value for whether you'd like to export threshold plots for each pass.}
}
\description{
The create_extra_parameters_list() function will allow users to modify any of the default
parameter values to allow for more specific linkage options.
}
\examples{
extra_params <- create_extra_parameters_list(output_linkage_iterations = TRUE, generate_linkage_report = TRUE)
}
