% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage_database_get_functions.R
\name{get_acceptence_thresholds}
\alias{get_acceptence_thresholds}
\title{Get Iteration Acceptance Threshold}
\usage{
get_acceptence_thresholds(linkage_db, iteration_id)
}
\arguments{
\item{linkage_db}{A database connection to the linkage metadata.}

\item{iteration_id}{An iteration number.}
}
\description{
The get_acceptence_thresholds() function will take in a linkage database connection
containing all the metadata, along with an iteration ID for the current iteration being run.
A list of key-value pairs are returned containing all the necessary thresholds for the
current iteration.
}
\examples{
sqlite_file <- file.choose() # Select the '.sqlite' linkage metadata file
linkage_db <- dbConnect(SQLite(), sqlite_file)
iteration_id <- 1
get_acceptence_thresholds(linkage_db, iteration_id)
}
