% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage_helper_functions.R
\name{get_algorithm_name}
\alias{get_algorithm_name}
\title{Get Iteration Implementation Name}
\usage{
get_algorithm_name(linkage_db, algorithm_id)
}
\arguments{
\item{linkage_db}{A database connection to the linkage metadata.}

\item{algorithm_id}{An algorithm number.}
}
\description{
The get_algorithm_name() function will take in a linkage database connection
containing all the metadata, along with an algorithm ID for the current algorithm being run.
A singular string is returned which is the user defined name of the the algorithm which
is currently being run.
}
\examples{
sqlite_file <- file.choose() # Select the '.sqlite' linkage metadata file
linkage_db <- dbConnect(SQLite(), sqlite_file)
algorithm_id <- 1
get_algorithm_name(linkage_db, algorithm_id)
}
