% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage_database_get_functions.R
\name{get_ground_truth_fields}
\alias{get_ground_truth_fields}
\title{Get Algorithm Ground Truth Keys}
\usage{
get_ground_truth_fields(linkage_db, algorithm_id)
}
\arguments{
\item{linkage_db}{A database connection to the linkage metadata.}

\item{algorithm_id}{An iteration number.}
}
\description{
The get_ground_truth_fields() function will take in a linkage database connection containing
all the metadata, along with an iteration ID for the current iteration being run.
A dataframe is returned which contains the left and right dataset fields + the linkage
rules that should be applied to those fields.
}
\examples{
sqlite_file <- file.choose() # Select the '.sqlite' linkage metadata file
linkage_db <- dbConnect(SQLite(), sqlite_file)
algorithm_id <- 1
get_ground_truth_fields(linkage_db, algorithm_id)
}
