% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage_database_get_functions.R
\name{get_implementation_name}
\alias{get_implementation_name}
\title{Get Iteration Implementation Name}
\usage{
get_implementation_name(linkage_db, iteration_id)
}
\arguments{
\item{linkage_db}{A database connection to the linkage metadata.}

\item{iteration_id}{An iteration number.}
}
\description{
The get_implementation_name() function will take in a linkage database connection
containing all the metadata, along with an iteration ID for the current iteration being run.
A singular string is returned which is what linkage implementation should be used by
the current running iteration.
}
\examples{
sqlite_file <- file.choose() # Select the '.sqlite' linkage metadata file
linkage_db <- dbConnect(SQLite(), sqlite_file)
iteration_id <- 1
get_implementation_name(linkage_db, iteration_id)
}
