% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage_helper_functions.R
\name{get_linkage_missingness_fields}
\alias{get_linkage_missingness_fields}
\title{Get Algorithm Missingness Fields}
\usage{
get_linkage_missingness_fields(linkage_db, algorithm_id)
}
\arguments{
\item{linkage_db}{A database connection to the linkage metadata.}

\item{algorithm_id}{An iteration number.}
}
\description{
The get_linkage_missingness_fields() function will take in a linkage database connection containing
all the metadata, along with an algorithm ID.
A dataframe is returned which contains the dataset fields to keep for gathering missing data indicators.
}
\examples{
sqlite_file <- file.choose() # Select the '.sqlite' linkage metadata file
linkage_db <- dbConnect(SQLite(), sqlite_file)
algorithm_id <- 1
get_linkage_output_fields(linkage_db, algorithm_id)
}
