% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage_helper_functions.R
\name{get_linkage_non_stratification_fields}
\alias{get_linkage_non_stratification_fields}
\title{Get Algorithm Non-Stratification Fields}
\usage{
get_linkage_non_stratification_fields(linkage_db, algorithm_id)
}
\arguments{
\item{linkage_db}{A database connection to the linkage metadata.}

\item{algorithm_id}{An algorithm ID}
}
\description{
The get_linkage_non_stratification_fields() function will take in a linkage database connection containing
all the metadata, along with an algorithm ID.
A dataframe is returned which contains the dataset fields of non-stratification field/column names.
}
\examples{
sqlite_file <- file.choose() # Select the '.sqlite' linkage metadata file
linkage_db <- dbConnect(SQLite(), sqlite_file)
algorithm_id <- 1
get_linkage_non_stratification_fields(linkage_db, algorithm_id)
}
