% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage_database_get_functions.R
\name{get_linkage_technique}
\alias{get_linkage_technique}
\title{Get Iteration Linkage Technique}
\usage{
get_linkage_technique(linkage_db, iteration_id)
}
\arguments{
\item{linkage_db}{A database connection to the linkage metadata.}

\item{iteration_id}{An iteration number.}
}
\description{
The get_linkage_technique() function will take in a linkage database connection
containing all the metadata, along with an iteration ID for the current iteration being run.
A singular string is returned which is what linkage technique should be used by
a specific linkage implementation.
}
\examples{
sqlite_file <- file.choose() # Select the '.sqlite' linkage metadata file
linkage_db <- dbConnect(SQLite(), sqlite_file)
iteration_id <- 1
get_linkage_technique(linkage_db, iteration_id)
}
