% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage_helper_functions.R
\name{get_linkage_technique_description}
\alias{get_linkage_technique_description}
\title{Get Iteration Linkage Technique Description}
\usage{
get_linkage_technique_description(linkage_db, iteration_id)
}
\arguments{
\item{linkage_db}{A database connection to the linkage metadata.}

\item{iteration_id}{An iteration number.}
}
\description{
The get_linkage_technique_description() function will take in a linkage database connection
containing all the metadata, along with an iteration ID for the current iteration being run.
A singular string is returned which is what the description of the linkage technique
being used during this pass.
}
\examples{
sqlite_file <- file.choose() # Select the '.sqlite' linkage metadata file
linkage_db <- dbConnect(SQLite(), sqlite_file)
iteration_id <- 1
get_linkage_technique_description(linkage_db, iteration_id)
}
