% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage_helper_functions.R
\name{get_linkage_time_trend_fields}
\alias{get_linkage_time_trend_fields}
\title{Get Linkage Time Trend Fields}
\usage{
get_linkage_time_trend_fields(linkage_db, dataset_code)
}
\arguments{
\item{linkage_db}{A database connection to the linkage metadata}

\item{dataset_code}{A character string matching an existing dataset code in the database}
}
\description{
The get_linkage_time_trend_fields() function will take in a linkage database connection
containing all the metadata, along with a dataset code pertaining to a specific dataset entry.
A list of both time trend variables is returned under the 'capture_month' and 'capture_year' variable names.
}
\examples{
sqlite_file <- file.choose() # Select the '.sqlite' linkage metadata file
linkage_db <- dbConnect(SQLite(), sqlite_file)
dataset_code <- "examplecode"
get_linkage_time_trend_fields(linkage_db, dataset_code)
}
