% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage_database_get_functions.R
\name{get_matching_keys}
\alias{get_matching_keys}
\title{Get Iteration Matching Keys}
\usage{
get_matching_keys(linkage_db, iteration_id)
}
\arguments{
\item{linkage_db}{A database connection to the linkage metadata.}

\item{iteration_id}{An iteration number.}
}
\description{
The get_matching_keys() function will take in a linkage database connection containing
all the metadata, along with an iteration ID for the current iteration being run.
A dataframe is returned which contains the left and right dataset fields + the linkage
rules that should be applied to those fields.
}
\examples{
sqlite_file <- file.choose() # Select the '.sqlite' linkage metadata file
linkage_db <- dbConnect(SQLite(), sqlite_file)
iteration_id <- 1
get_matching_keys(linkage_db, iteration_id)
}
