% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage_helper_functions.R
\name{load_linkage_file}
\alias{load_linkage_file}
\title{Load Linkage File}
\usage{
load_linkage_file(dataset_file)
}
\arguments{
\item{dataset_file}{A path to the input file}
}
\description{
The load_linkage_file() function will take in a dataset file of a specific type, which may
be csv, txt, sqlite, sas7bdat, xlsx, and will attempt to load and return it.
}
\examples{
file_path <- choose.file() # Select an input file
load_linkage_file(file_path)
}
