% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_linkage_scripts.R
\name{run_main_linkage}
\alias{run_main_linkage}
\title{Run Data Linkage Function}
\usage{
run_main_linkage(
  left_dataset_file,
  right_dataset_file,
  linkage_metadata_file,
  algorithm_ids,
  extra_parameters,
  progress_callback = NULL
)
}
\arguments{
\item{left_dataset_file}{The left dataset file to be linked to the right dataset file.}

\item{right_dataset_file}{The right dataset file that will be be linked with the left dataset file.}

\item{linkage_metadata_file}{The .SQLITE file of metadata information used for data linkage.}

\item{algorithm_ids}{A vector of integers which are the algorithm IDs that will be ran.}

\item{extra_parameters}{A list of optional user defined parameters that affects output.}

\item{progress_callback}{An optional parameter which is a SHINY UI object passed to provide UI users with a progress bar of how far along completion is.}
}
\value{
A list of linkage result data if the user specified in the extra parameters, otherwise NULL is returned
}
\description{
The run_main_linkage() function will take in a left and right dataset, using the linkage metadata file
to determine which linkage classes to use for each iteration of an algorithm, with extra parameters to
define what kinds of outputs are wanted.
}
\examples{
left_dataset_file   <- file.choose()
right_dataset_file  <- file.choose()
linkage_sqlite_file <- file.choose()
algorithm_ids       <- c(1, 2, 3)
extra_parameters    <- create_extra_parameters_list(linkage_output_folder = choose.dir(), linkage_report_type = 2)
results             <- run_main_linkage(left_dataset_file, right_dataset_file, linkage_sqlite_file, algorithm_ids, extra_parameters)
}
