% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage_metadata_ui.R
\name{start_linkage_metadata_ui}
\alias{start_linkage_metadata_ui}
\title{Start Linkage Metadata UI}
\usage{
start_linkage_metadata_ui(metadata_file_path, username)
}
\description{
Upon call, and passing a valid linkage metadata file, the application will run using the supplied metadata file
as the file that can be modified. This involves viewing entries, adding new entires, modifying existing entires
or deleting them.
}
\examples{
my_metadata_file <- file.choose() # Select .sqlite file
my_username <- "John Doe"
start_linkage_metadata_ui(my_metadata_file, my_username)
}
