% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage_helper_functions.R
\name{create_extra_parameters_list}
\alias{create_extra_parameters_list}
\title{Create Extra Parameters List}
\usage{
create_extra_parameters_list(
  linkage_output_folder = NULL,
  output_linkage_iterations = FALSE,
  generate_linkage_report = FALSE,
  data_linker = NULL,
  standardize_names_file_path = NULL
)
}
\arguments{
\item{linkage_output_folder}{A directory for which linkage output files will be written to.}

\item{output_linkage_iterations}{A TRUE or FALSE value for whether you'd like each iteration to write the linked pairs (UNEDITED) to the output directory.}

\item{generate_linkage_report}{A TRUE or FALSE value for whether you'd like a Linkage Quality Report to be generated and written to the output directory (from the linkrep package).}

\item{data_linker}{A single string input for whom performed the data linkage (used for generating a Linkage Quality Report).}

\item{standardize_names_file_path}{A path to a CSV containing common alternative spellings of names that will standardize to a singular spelling. Must have the columns 'START' and 'LABEL'.}
}
\description{
The create_extra_parameters_list() function will allow users to modify any of the default
parameter values to allow for more specific linkage options.
}
\examples{
extra_params <- create_extra_parameters_list(output_linkage_iterations = TRUE, generate_linkage_report = TRUE)
}
