% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage_database_get_functions.R
\name{get_standardized_names}
\alias{get_standardized_names}
\title{Get Standardized Names}
\usage{
get_standardized_names(
  file_path,
  data_field,
  lookupvector = common_standardized_names
)
}
\arguments{
\item{file_path}{A path to the .csv file that will be read}

\item{data_field}{A vector of names that will be standardized.}

\item{lookupvector}{The vector of names we'll try to unname.}
}
\description{
The get_standardized_names() function will take in a standardizing names file in
the form of the csv which must contain the columns 'START' which indicates all possible
forms of a common name, and a column 'LABEL' which is the most common spelling of that
name which it will be standardized to. If no spelling is found, the name is replaced with
NA
}
\examples{
file_path <- choose.file() # Select the '.csv'
data_field <- c("John", "Johnnie", "Johnny", "Jon")
get_standardized_names(file_path, data_field)
}
